Refa：migrate agent webhook routes to REST APIs
