+++ web/src/pages/agent/hooks/use-build-webhook-url.ts
@@ -3,6 +3,6 @@ import { useParams } from 'react-router';
 export function useBuildWebhookUrl() {
   const { id } = useParams();
 
-  const text = `${location.protocol}//${location.host}/api/v1/webhook/${id}`;
+  const text = `${location.protocol}//${location.host}/api/v1/agents/${id}/webhook`;
   return text;
 }
